fix: :bug: 최초 실행 시 마커가 등록되지 않는 버그 수정(#8)
